fixed warnings

+++ src/filters/isoffin_load.c
@@ -443,7 +443,7 @@ void isor_declare_objects(ISOMReader *read)
 		if (sample_count && w && h) {
 			u64 mdur = gf_isom_get_media_duration(read->mov, i+1);
 			mdur /= sample_count;
-			gf_filter_pid_set_property(pid, GF_PROP_PID_FPS, &PROP_FRAC_INT(ch->time_scale, mdur));
+			gf_filter_pid_set_property(pid, GF_PROP_PID_FPS, &PROP_FRAC_INT(ch->time_scale, (u32) mdur));
 		}
 
 		track_dur = (Double) (s64) ch->duration;
